+++ inst/testScripts/BNR/testStepDown.R
@@ -43,7 +43,7 @@ testStepDown <- function(m, rho, B, pi0, SNR, alpha, flavor=c("equi", "Mein2006"
     resMat <- NULL
     for (refFam in c("kFWER", "Simes")) {
         ## Step-down control
-        res <- jointFWERControl(X0, refFamily=refFam, alpha=alpha, stat=x)
+        res <- jointFWERControl(X0, refFamily=refFam, alpha=alpha, stat=x, verbose=FALSE)
         thrMat <- res$stepsDown$thr
         nSteps <- ncol(thrMat)
         thr0 <- thrMat[, 1]
@@ -53,13 +53,13 @@ testStepDown <- function(m, rho, B, pi0, SNR, alpha, flavor=c("equi", "Mein2006"
         ## comparison with *Oracle step-down* JFWER thresholds (only the step-down is Oracle)
         xx <- rep(Inf, length(x))
         xx[H0] <- -Inf
-        resOracle <- jointFWERControl(X0, refFamily=refFam, alpha=alpha, stat=xx)
+        resOracle <- jointFWERControl(X0, refFamily=refFam, alpha=alpha, stat=xx, verbose=FALSE)
         thrO <- resOracle$thr
 
         ## *Oracle JFWER* thresholds (double Oracle!!!)
         X0Oracle <- X0[H0, ]
         resOracleJ <- jointFWERControl(X0Oracle, refFamily=refFam, alpha=alpha,
-                                       maxStepsDown=0)  ## single step
+                                       maxStepsDown=0, verbose=FALSE)  ## single step
         thrOJ <- c(resOracleJ$thr, rep(-Inf, m1))
 
         resFam <- cbind("0"=thr0, "SD"=thrSD, "Oracle"=thrO, "Oracle2"=thrOJ)
@@ -79,9 +79,20 @@ testStepDown <- function(m, rho, B, pi0, SNR, alpha, flavor=c("equi", "Mein2006"
     } else {     ## summarize results
         x0 <- x[which(H==0)]
         x1 <- x[which(H==1)]
-        rej0 <- apply(resMat, 2, rej, x0)
-        rej1 <- apply(resMat, 2, rej, x1)
-        res <- rbind(rej0, rej1)
+        rej0 <- apply(resMat, 2, rej, x0) ## nb of k s.t. |Rk \cup H0| > k-1 ('violators')
+        rej1 <- apply(resMat, 2, rej, x1) ## nb of k s.t. |Rk \cup H1| > k-1 ('good catchers')
+
+        rejk0 <- apply(resMat, 2, rejk, x0)
+        rejk1 <- apply(resMat, 2, rejk, x1)
+
+        ## sanity checks
+        stopifnot(all.equal(colSums(rejk0>=0), rej0))
+        stopifnot(all.equal(colSums(rejk1>=0), rej1))
+
+        v0 <- (m0-1-pmax(0, matrixStats::colMaxs(rejk0)))/m0  ## estimate of Vbar(H0)/m0
+        s1 <- pmax(0, matrixStats::colMaxs(rejk1))/m1         ## estimate of Sbar(H1)/m1
+
+        res <- rbind(rej0, rej1, v0, s1)
     }
     return(res)
 }
@@ -92,3 +103,8 @@ rej <- function(thr, x) {
     sum(nover >= 1:length(thr))
 }
 
+## to estimate JFWER and power
+rejk <- function(thr, x) {
+    nover <- sapply(thr, FUN=function(ss) sum(x > ss))
+    nover-1:length(thr)
+}